@@ -1,6 +1,6 @@
 {
   "name": "filtered-input",
-  "version": "1.2.2",
+  "version": "1.2.3",
   "description": "A simple input that only accepts characters based on a Regex pattern",
   "main": "dist/filtered-input.js",
   "repository": {
@@ -31,16 +31,17 @@
     "babel-plugin-transform-object-rest-spread": "^6.20.2",
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
-    "eslint": "^3.12.2",
+    "eslint": "^3.19.0",
     "eslint-loader": "^1.6.1",
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-jsx-a11y": "^3.0.2",
     "eslint-plugin-react": "^6.8.0",
     "jest": "^18.0.0",
     "numeral": "^2.0.4",
-    "react": "^15.4.1",
-    "react-dom": "^15.4.1",
-    "react-test-renderer": "^15.4.1",
+    "prop-types": "^15.5.7",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-test-renderer": "^15.5.4",
     "webpack": "^1.14.0"
   }
 }
